+++ HISTORY.rst
@@ -62,3 +62,4 @@ History
 4.0.0 (2020-10-xx)
 ------------------
 * Python 3.7 supported
+* Pagination with continue parameter over linked resources supported